backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
Pillow==10.4.0
python-multipart==0.0.9
reportlab==4.2.0
